Initial commit.
There’s a demo; it’s missing proper persistence and utility,
and there should also be a listener for the `move` event, to
provide proper input.

The slider and buttons also need to be moved directly beneath
the crop box, and the slider needs to be replaced, because its
looks are inconsistent and it’s hard to style at the moment…
though vendor prefixes *are* more attractive than more JavaScript.

Documentation and more comprehensive testing to follow.
Opera Mini almost certainly doesn’t work properly with Cut.